pyproject.toml
[tool.poetry]
name = "open-llm-on-lab"
version = "0.1.0"
description = ""
authors = ["nozomu-karai <karai.nozomu.77c@st.kyoto-u.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9"
pytorch-lightning = "^2.0.9"
transformers = "^4.33.2"
hydra-core = "^1.3.2"
torch = "1.13.0"
accelerate = "^0.23.0"
bitsandbytes = "^0.41.1"
scipy = "1.6.0"
optimum = "^1.13.2"
auto-gptq = "^0.4.2"
humanize = "^4.8.0"
gputil = "^1.4.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.9.1"
isort = "^5.12.0"
pre-commit = "^3.4.0"

[tool.black]
line-length = 120
target-version = ["py37"]

[tool.isort]
line_length = 120

[tool.mypy]
python_version = 3.8

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
